+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/preview/CarRouteLine.kt
@@ -3,6 +3,7 @@ package com.mapbox.androidauto.car.preview
 import com.mapbox.androidauto.car.MainCarContext
 import com.mapbox.androidauto.car.internal.extensions.getStyle
 import com.mapbox.androidauto.car.internal.extensions.handleStyleOnAttached
+import com.mapbox.androidauto.car.internal.extensions.handleStyleOnDetached
 import com.mapbox.androidauto.car.routes.NavigationRoutesProvider
 import com.mapbox.androidauto.car.routes.RoutesListener
 import com.mapbox.androidauto.car.routes.RoutesProvider
@@ -25,6 +26,12 @@ import com.mapbox.navigation.ui.maps.route.line.model.MapboxRouteLineOptions
 import com.mapbox.navigation.ui.maps.route.line.model.NavigationRouteLine
 import com.mapbox.navigation.ui.maps.route.line.model.RouteLineColorResources
 import com.mapbox.navigation.ui.maps.route.line.model.RouteLineResources
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.MainScope
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.cancelChildren
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
 
 /**
  * This class is to simplify the interaction with [MapboxRouteLineApi], [MapboxRouteArrowView]
@@ -52,6 +59,8 @@ class CarRouteLine internal constructor(
     private lateinit var routeLineApi: MapboxRouteLineApi
     private lateinit var routeArrowApi: MapboxRouteArrowApi
     private lateinit var routeArrowView: MapboxRouteArrowView
+
+    private var styleScope: CoroutineScope? = null
 
     private val onPositionChangedListener = OnIndicatorPositionChangedListener { point ->
         val result = routeLineApi.updateTraveledRouteLine(point)
@@ -98,6 +107,7 @@ class CarRouteLine internal constructor(
     override fun onAttached(mapboxCarMapSurface: MapboxCarMapSurface) {
         logAndroidAuto("CarRouteLine carMapSurface loaded $mapboxCarMapSurface")
         val locationPlugin = mapboxCarMapSurface.mapSurface.location
+        val scope = MainScope().also { styleScope = it }
         styleLoadedListener = mapboxCarMapSurface.handleStyleOnAttached { style ->
             val routeLineOptions = getMapboxRouteLineOptions(style)
             routeLineView = MapboxRouteLineView(routeLineOptions)
@@ -114,15 +124,29 @@ class CarRouteLine internal constructor(
             locationPlugin.addOnIndicatorPositionChangedListener(onPositionChangedListener)
             mainCarContext.mapboxNavigation.registerRouteProgressObserver(routeProgressObserver)
             routesProvider.registerRoutesListener(routesListener)
+
+            scope.coroutineContext.cancelChildren()
+            scope.launch {
+                mainCarContext.routeAlternativesEnabled.collect { enabled ->
+                    if (enabled) {
+                        routeLineView.showAlternativeRoutes(style)
+                    } else {
+                        routeLineView.hideAlternativeRoutes(style)
+                    }
+                }
+            }
         }
     }
 
     override fun onDetached(mapboxCarMapSurface: MapboxCarMapSurface) {
         logAndroidAuto("CarRouteLine carMapSurface detached $mapboxCarMapSurface")
         val mapSurface = mapboxCarMapSurface.mapSurface
+        mapboxCarMapSurface.handleStyleOnDetached(styleLoadedListener)
         mapSurface.location.removeOnIndicatorPositionChangedListener(onPositionChangedListener)
         mainCarContext.mapboxNavigation.unregisterRouteProgressObserver(routeProgressObserver)
         routesProvider.unregisterRoutesListener(routesListener)
+        styleScope?.cancel()
+        styleScope = null
     }
 
     private fun getMapboxRouteLineOptions(style: Style): MapboxRouteLineOptions {